Add files via upload
Skeleton Code to get us going in the right direction. All position values (textfields, buttons, etc..) are not set as I can't run it. I'm getting a consistent error that I would most likely need to go to office hours to fix. This is a text file but should have no JavaFx errors upon pasting it into eclipse if you have it set up right.
Download it and check it out